Permissions

+++ app/Http/Controllers/Admin/ReservationController.php
@@ -21,8 +21,13 @@ class ReservationController extends Controller {
 	 */
 	public function index()
 	{
-		$reservations = SeatReservation::all();
-		return view('seating.reservation.index')->withReservations($reservations);
+		if (Sentinel::getUser()->hasAccess(['admin.reservation.*'])) {
+			$reservations = SeatReservation::all();
+			return view('seating.reservation.index')->withReservations($reservations);
+		} else {
+			return Redirect::back()->with('messagetype', 'warning')
+								->with('message', 'You do not have access to this page!');
+		}
 	}
 
 	/**
@@ -64,8 +69,13 @@ class ReservationController extends Controller {
 	 */
 	public function edit($id)
 	{
-		$reservation = SeatReservation::find($id);
-		return view('seating.reservation.edit')->withReservation($reservation);
+		if (Sentinel::getUser()->hasAccess(['admin.reservation.update'])) {
+			$reservation = SeatReservation::find($id);
+			return view('seating.reservation.edit')->withReservation($reservation);
+		} else {
+			return Redirect::back()->with('messagetype', 'warning')
+								->with('message', 'You do not have access to this page!');
+		}
 	}
 
 	/**
@@ -76,39 +86,45 @@ class ReservationController extends Controller {
 	 */
 	public function update($id, ReservationEditRequest $request)
 	{
-		$reservation 					= SeatReservation::find($id);
-
-		$updateseat						= Seats::find($reservation->seat->id);
-		$updateseat->reservation_id		= $reservation->id;
-		$updateseat->save();
-
-		if($reservation->status_id == 1) { // 1 = Reserved, 2 = Temporary Reserved
+		if (Sentinel::getUser()->hasAccess(['admin.reservation.update'])) {
 			
-			if($reservation->ticket) {
-				$reservation->ticket->delete();
+			$reservation 					= SeatReservation::find($id);
+
+			$updateseat						= Seats::find($reservation->seat->id);
+			$updateseat->reservation_id		= $reservation->id;
+			$updateseat->save();
+
+			if($reservation->status_id == 1) { // 1 = Reserved, 2 = Temporary Reserved
+				
+				if($reservation->ticket) {
+					$reservation->ticket->delete();
+				}
+
+				$seatticket 					= new SeatTicket;
+				$seatticket->barcode 			= mt_rand(1000000000, 2147483647);
+				$seatticket->reservation_id		= $reservation->id;
+				$seatticket->user_id			= $request->get('reservedfor_id');
+				$seatticket->save();
+
+				$reservation->ticket_id 		= $seatticket->id;
+
 			}
 
-			$seatticket 					= new SeatTicket;
-			$seatticket->barcode 			= mt_rand(1000000000, 2147483647);
-			$seatticket->reservation_id		= $reservation->id;
-			$seatticket->user_id			= $request->get('reservedfor_id');
-			$seatticket->save();
+			$reservation->reservedfor_id 	= $request->get('reservedfor_id');
+			$reservation->seat_id 			= $request->get('seat_id');
 
-			$reservation->ticket_id 		= $seatticket->id;
-
-		}
-
-		$reservation->reservedfor_id 	= $request->get('reservedfor_id');
-		$reservation->seat_id 			= $request->get('seat_id');
-
-		if($reservation->save()) {
-			return Redirect::route('admin-seating-reservation-edit', $id)
-					->with('messagetype', 'success')
-					->with('message', 'The reservation has now been saved!');
+			if($reservation->save()) {
+				return Redirect::route('admin-seating-reservation-edit', $id)
+						->with('messagetype', 'success')
+						->with('message', 'The reservation has now been saved!');
+			} else {
+				return Redirect::route('admin-seating-reservation-edit', $id)
+					->with('messagetype', 'danger')
+					->with('message', 'Something went wrong while saving the reservation.');
+			}
 		} else {
-			return Redirect::route('admin-seating-reservation-edit', $id)
-				->with('messagetype', 'danger')
-				->with('message', 'Something went wrong while saving the reservation.');
+			return Redirect::back()->with('messagetype', 'warning')
+								->with('message', 'You do not have access to this page!');
 		}
 	}
